Add 'Configuration#detachAll'
For each unique prefix in the configuration, it produces a new
configuration containing the values whose keys have the common
prefix. It returns a map from prefixes to configurations.

This is useful in the following scenario:

  accounts {
    alice {
      group: admins
      name:  Alice
    }
    bob {
      group: users
      name:  Bob
    }
  }

Here the account identifiers, 'alice' and 'bob', should obviously
not be hard-coded into the application. 'Configuration#detachAll'
on the 'accounts' sub-configuration provides a convenient discovery
mechanism for the configured accounts.