Upgrade to Scala 2.10 (at the cost of losing 2.8 compatibility)
Because of the ScalaTest library dependence, support for older than 2.9.2
version is dropped in this commit, because ScalaTest is not built for older
than Scala 2.9.2.

There would definitely be ways of cross-compiling Configrity for both Scala
2.10, 2.9 and 2.8, as nothing in the code needs to be changed to be compatible
with 2.10.

You might not want to pull this in as-is (because of the lost compatibility),
but it does show that Configrity works just fine with 2.10 and it would be
beneficial to get an artifact published for that as well.

Signed-off-by: Aki Saarinen <aki@akisaarinen.fi>